handler.go に Handler構造体(とそのmethod)を作成
